refactor: Rush 이벤트 구조 수정
